# src/docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Bridge-LFU API",
    "version": "1.0.0",
    "description": "Bridge LFU (Licence Follow-Up) — API de suivi des licences logicielles et de l'obsolescence des équipements pour les clients de Bridge.\n\nFonctionnalités principales:\n- Authentification et gestion des utilisateurs (rôles: admin, technicien, client)\n- Gestion des clients, licences, équipements, fournisseurs, types\n- Notifications et alertes (échéances licences, obsolescence équipements)\n- Tableaux de bord et statistiques\n- Rapports exportables (JSON/CSV/PDF/Excel)\n\nConventions d’API:\n- Authentification: session via Supabase (cookies). Lancez d'abord POST /api/auth/login dans le même navigateur/onglet.\n- Pagination: paramètres `page` et `limit`, réponses standard `{ data, count, page, totalPages, hasMore }`.\n- Filtres: exposés par ressource (ex: `status`, `client_id`, `type_id`, `date_from`, `date_to`).\n- Exports: utilisez le paramètre `format` (json|csv|pdf|excel) sur les endpoints de rapports.\n\nSupport: contactez l'équipe Bridge pour l’onboarding et les accès."
  },
  "servers": [
    { "url": "/", "description": "Serveur courant" }
  ],
...
  "paths": {
...
    "/api/equipment/export": { "get": { "tags": ["Equipment"], "summary": "Exporter les équipements", "parameters": [ { "in": "query", "name": "format", "schema": { "type": "string", "enum": ["csv", "xlsx", "pdf"], "default": "csv" } } ], "responses": { "200": { "description": "Fichier export" } } } },
...
  },
...
}